+++ src/com/kstenschke/referencer/referencers/insertOrCopy/InsertOrCopyReferencer.java
@@ -31,6 +31,7 @@ import com.kstenschke.referencer.utils.UtilsFile;
 import com.kstenschke.referencer.utils.UtilsString;
 import org.apache.commons.lang.StringUtils;
 
+import javax.annotation.Nullable;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -42,91 +43,84 @@ public class InsertOrCopyReferencer {
      * @param e Action system event
      * @return Items of the events context
      */
-    public static String[] getItems(AnActionEvent e) {
-        String[] referencesArr;
-
+    @Nullable public static String[] getItems(AnActionEvent e) {
         final Project project = e.getData(PlatformDataKeys.PROJECT);
         Editor editor = e.getData(PlatformDataKeys.EDITOR);
 
-        if (project != null && editor != null) {
-            final Document document = editor.getDocument();
-
-            // File path and name
-            VirtualFile file = FileDocumentManager.getInstance().getFile(document);
-            String filePath = (file != null) ? file.getPath() : "";
-            String fileExtension = UtilsFile.getExtensionByDocument(document);
-
-            // Get line number the caret is in
-            int caretOffset = editor.getCaretModel().getOffset();
-
-            SelectionModel selectionModel = editor.getSelectionModel();
-            int selStart = selectionModel.getSelectionStart();
-            int selEnd = selectionModel.getSelectionEnd();
-
-            int lineSelStart = document.getLineNumber(selStart);
-            int lineSelEnd = document.getLineNumber(selEnd);
-
-            // Grab the "word" to the left of the caret - all java identifier characters
-            String wordLeftOfCaret = UtilsString.getWordLeftOfOffset(document.getCharsSequence(), caretOffset);
-            // Grab the "string" to the left of the caret - all non white-space characters
-            String stringLeftOfCaret = UtilsString.getStringLeftOfOffset(document.getCharsSequence(), caretOffset);
-
-            // Setup list of items
-            List<String> referenceItems = new ArrayList<>();
-
-            // Add date/timestamps
-            referenceItems.add(StaticTexts.POPUP_SECTION_TITLE_DATE_TIME);
-            referenceItems.addAll(InsertOrCopyReferencerDateTime.getReferenceItems());
-
-            // Add file / path items
-            referenceItems.add(StaticTexts.POPUP_SECTION_TITLE_FILES_PATHS);
-            referenceItems.addAll(InsertOrCopyReferencerFilesFolders.getReferenceItems(e));
-
-            // Add selection info
-            if (selectionModel.hasSelection() && lineSelEnd > lineSelStart) {
-                referenceItems.add(filePath + " / Selection: " + (lineSelStart + 1) + " - " + (lineSelEnd + 1));
-            }
-
-            // Add JavaScript items
-            if (UtilsFile.isJavaScriptFileExtension(fileExtension)) {
-                referenceItems.add(StaticTexts.POPUP_SECTION_TITLE_JAVASCRIPT);
-                referenceItems.addAll(InsertOrCopyReferencerJavascript.getReferenceItems(e));
-            } else if (UtilsFile.isPhpFileExtension(fileExtension)) {
-                // Add PHP items
-                referenceItems.add(StaticTexts.POPUP_SECTION_TITLE_PHP);
-                referenceItems.addAll(InsertOrCopyReferencerPhp.getReferenceItems(e));
-            }
-
-            // Add all line-parts in current document that begin the same as the word at the caret
-            if (wordLeftOfCaret != null && !wordLeftOfCaret.isEmpty() && wordLeftOfCaret.length() > 1
-             || (stringLeftOfCaret != null && !stringLeftOfCaret.isEmpty() && stringLeftOfCaret.length() > 2)
-            ) {
-                String docText = document.getText();
-                String[] wordLineParts = null;
-                if (wordLeftOfCaret != null && !wordLeftOfCaret.isEmpty()) {
-                    wordLineParts = docText.split(wordLeftOfCaret);
-                }
-
-                String[] allLineParts;
-                if (stringLeftOfCaret != null && !stringLeftOfCaret.isEmpty()) {
-                    String[] stringLineParts = docText.split(Pattern.quote(stringLeftOfCaret));
-                    allLineParts = UtilsArray.merge(wordLineParts, stringLineParts);
-                } else {
-                    allLineParts = wordLineParts;
-                }
-
-                if (allLineParts != null && allLineParts.length > 0) {
-                    List<String> listLineParts = Arrays.asList(allLineParts);
-                    addReferenceItems(referenceItems, listLineParts);
-                }
-            }
-
-            referencesArr = referenceItems.toArray(new String[referenceItems.size()]);
-        } else {
-            referencesArr = null;
+        if (project == null || editor == null) {
+            return null;
         }
 
-        return referencesArr;
+        final Document document = editor.getDocument();
+
+        /* File path and name */
+        VirtualFile file = FileDocumentManager.getInstance().getFile(document);
+        String filePath = (file != null) ? file.getPath() : "";
+        String fileExtension = UtilsFile.getExtensionByDocument(document);
+
+        int caretOffset = editor.getCaretModel().getOffset();                   /* Get line number the caret is in */
+
+        SelectionModel selectionModel = editor.getSelectionModel();
+        int selStart = selectionModel.getSelectionStart();
+        int selEnd = selectionModel.getSelectionEnd();
+
+        int lineSelStart = document.getLineNumber(selStart);
+        int lineSelEnd = document.getLineNumber(selEnd);
+
+        /* Grab the "word" to the left of the caret - all java identifier characters */
+        String wordLeftOfCaret = UtilsString.getWordLeftOfOffset(document.getCharsSequence(), caretOffset);
+        /* Grab the "string" to the left of the caret - all non white-space characters */
+        String stringLeftOfCaret = UtilsString.getStringLeftOfOffset(document.getCharsSequence(), caretOffset);
+
+        /* Setup list of items */
+        List<String> referenceItems = new ArrayList<>();
+
+        /* Add date/timestamps */
+        referenceItems.add(StaticTexts.POPUP_SECTION_TITLE_DATE_TIME);
+        referenceItems.addAll(InsertOrCopyReferencerDateTime.getReferenceItems());
+
+        /* Add file / path items */
+        referenceItems.add(StaticTexts.POPUP_SECTION_TITLE_FILES_PATHS);
+        referenceItems.addAll(InsertOrCopyReferencerFilesFolders.getReferenceItems(e));
+
+        /* Add selection info */
+        if (selectionModel.hasSelection() && lineSelEnd > lineSelStart) {
+            referenceItems.add(filePath + " / Selection: " + (lineSelStart + 1) + " - " + (lineSelEnd + 1));
+        }
+
+        if (UtilsFile.isJavaScriptFileExtension(fileExtension)) {                   /* Add JavaScript items */
+            referenceItems.add(StaticTexts.POPUP_SECTION_TITLE_JAVASCRIPT);
+            referenceItems.addAll(InsertOrCopyReferencerJavascript.getReferenceItems(e));
+        } else if (UtilsFile.isPhpFileExtension(fileExtension)) {                   /* Add PHP items */
+            referenceItems.add(StaticTexts.POPUP_SECTION_TITLE_PHP);
+            referenceItems.addAll(InsertOrCopyReferencerPhp.getReferenceItems(e));
+        }
+
+        /* Add all line-parts in current document that begin the same as the word at the caret */
+        if (wordLeftOfCaret != null && !wordLeftOfCaret.isEmpty() && wordLeftOfCaret.length() > 1
+         || (stringLeftOfCaret != null && !stringLeftOfCaret.isEmpty() && stringLeftOfCaret.length() > 2)
+        ) {
+            String docText = document.getText();
+            String[] wordLineParts = null;
+            if (wordLeftOfCaret != null && !wordLeftOfCaret.isEmpty()) {
+                wordLineParts = docText.split(wordLeftOfCaret);
+            }
+
+            String[] allLineParts;
+            if (stringLeftOfCaret != null && !stringLeftOfCaret.isEmpty()) {
+                String[] stringLineParts = docText.split(Pattern.quote(stringLeftOfCaret));
+                allLineParts = UtilsArray.merge(wordLineParts, stringLineParts);
+            } else {
+                allLineParts = wordLineParts;
+            }
+
+            if (allLineParts != null && allLineParts.length > 0) {
+                List<String> listLineParts = Arrays.asList(allLineParts);
+                addReferenceItems(referenceItems, listLineParts);
+            }
+        }
+
+        return referenceItems.toArray(new String[referenceItems.size()]);
     }
 
     private static void addReferenceItems(List<String> referenceItems, List<String> listLineParts) {