+++ src/com/kstenschke/referencer/resources/StaticTexts.java
@@ -21,16 +21,16 @@ public class StaticTexts {
 
     @NonNls public static final String SETTINGS_DISPLAY_NAME = "Referencer";
 
-    // Popup titles
+    /* Popup titles */
     @NonNls public static final String POPUP_TITLE_ACTION_COPY = "Select what to copy";
     @NonNls public static final String POPUP_TITLE_ACTION_INSERT = "Select Insertion";
     @NonNls public static final String POPUP_TITLE_ACTION_GO = "Select where to go";
 
-    // Popup items
+    /* Popup items */
     @NonNls public static final String POPUP_ITEM_METHODS_IN_FILE = "List of methods in current file";
     public static final String POPUP_ITEM_OPEN_FILES = "List of currently opened files";
 
-    // Popup section titles
+    /* Popup section titles */
     @NonNls public static final String POPUP_ITEM_PREFIX_SECTION_TITLE = "SECTIONTITLE:";
     @NonNls public static final String POPUP_SECTION_TITLE_DATE_TIME = "SECTIONTITLE: Date / Time";
     @NonNls public static final String POPUP_SECTION_TITLE_FILES_PATHS = "SECTIONTITLE: Files / Paths";
@@ -41,9 +41,9 @@ public class StaticTexts {
     @NonNls public static final String POPUP_SECTION_BOOKMARKS = "SECTIONTITLE: Bookmarks";
     @NonNls public static final String POPUP_SECTION_FUNCTIONS = "SECTIONTITLE: Methods";
 
-    // Go action context menu
+    /* Go action context menu */
     @NonNls public static final String POPUP_GO_REMOVE_ALL_BOOKMARKS = "Remove all Bookmarks from this File";
 
-    // Notifications
+    /* Notifications */
     @NonNls public static final String NOTIFY_GOTO_NONE_FOUND = "No \"Go To...\" destinations found";
 }